requirements-pip.txt
adbutils==2.7.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioquic==1.2.0
aiosignal==1.3.1
albucore==0.0.13
albumentations==1.4.13
annotated-types==0.7.0
anyio==4.4.0
apkutils2==1.0.0
asgiref==3.8.1
astor==0.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cigam==0.0.3
click==8.1.7
cryptography==43.0.0
cython==3.0.11
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
editorconfig==0.12.4
eval-type-backport==0.2.0
filelock==3.15.4
fire==0.6.0
flask==3.0.2
frozenlist==1.4.1
fsspec==2024.9.0
gdown==5.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
imageio==2.34.2
imgaug==0.4.0
itsdangerous==2.2.0
jinja2==3.1.4
jsbeautifier==1.15.1
kaitaistruct==0.10
lazy-loader==0.4
ldap3==2.9.1
linkify-it-py==2.0.3
lmdb==1.5.1
lxml==5.3.0
markdown-it-py==3.0.0
markupsafe==2.1.5
mdit-py-plugins==0.4.1
mdurl==0.1.2
mitmproxy==10.4.2
mitmproxy-rs==0.6.3
mobly==1.12.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.1.0
openai==1.37.1
opencv-contrib-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.8.1
paddlepaddle==2.6.2
passlib==1.7.4
portpicker==1.6.0
protobuf==5.27.3
publicsuffix2==2.20191221
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.8.2
pydantic-core==2.20.1
pyelftools==0.31
pygments==2.18.0
pylsqpack==0.3.18
pymupdf==1.24.0
pyopenssl==24.2.1
pyparsing==3.1.2
pyperclip==1.9.0
pysocks==1.7.1
python-docx==1.1.2
python-dotenv==1.0.1
python-levenshtein==0.26.0
pytz==2024.1
pyyaml==6.0.1
rapidfuzz==3.9.7
recoverpy==2.1.9
retry==0.9.2
rich==13.8.0
ruamel-yaml==0.18.6
ruamel-yaml-clib==0.2.8
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
snippet-uiautomator==1.1.2
sortedcontainers==2.4.0
soupsieve==2.5
supervision==0.22.0
termcolor==2.4.0
textual==0.72.0
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tzdata==2024.1
uc-micro-py==1.0.3
uiautomator2==3.2.2
ultralytics==8.2.96
ultralytics-thop==2.0.6
urwid==2.6.15
waitress==3.0.0
wcwidth==0.2.13
werkzeug==3.0.3
wsproto==1.2.0
xmltodict==0.13.0
yarl==1.11.1
zstandard==0.23.0

